MasterclassTestingMachineLearning/requirements.txt
acres>=0.3.0
aiohappyeyeballs>=2.5.0
aiohttp>=3.11.13
aiosignal>=1.3.2
annotated-types>=0.7.0
anyio>=4.8.0
appdirs>=1.4.4
attrs>=25.1.0
certifi>=2025.1.31
charset-normalizer>=3.4.1
ci-info>=0.3.0
click>=8.1.8
colorama>=0.4.6
configobj>=5.0.9
configparser>=7.2.0
dataclasses-json>=0.6.7
datasets>=3.3.2
dill>=0.3.8
diskcache>=5.6.3
distro>=1.9.0
etelemetry>=0.3.1
evaluate>=0.4.3
faiss-cpu>=1.10.0
filelock>=3.17.0
fitz>=0.0.1.dev2
frozenlist>=1.5.0
fsspec>=2024.12.0
greenlet>=3.1.1
h11>=0.14.0
httpcore>=1.0.7
httplib2>=0.22.0
httpx>=0.28.1
httpx-sse>=0.4.0
huggingface-hub>=0.29.2
idna>=3.10
isodate>=0.6.1
jiter>=0.8.2
jsonpatch>=1.33
jsonpointer>=3.0.0
langchain>=0.3.20
langchain-community>=0.3.19
langchain-core>=0.3.43
langchain-openai>=0.3.8
langchain-text-splitters>=0.3.6
langsmith>=0.3.13
looseversion>=1.3.0
lxml>=5.3.1
marshmallow>=3.26.1
multidict>=6.1.0
multiprocess>=0.70.16
mypy-extensions>=1.0.0
nest-asyncio>=1.6.0
networkx>=3.4.2
nibabel>=5.3.2
nipype>=1.9.2
numpy>=2.2.3
openai>=1.65.5
orjson>=3.10.15
packaging>=24.2
pandas>=2.2.3
pathlib>=1.0.1
propcache>=0.3.0
prov>=2.0.1
puremagic>=1.28
pyarrow>=19.0.1
pydantic>=2.10.6
pydantic-settings>=2.8.1
pydantic_core>=2.27.2
pydot>=3.0.4
PyMuPDF>=1.25.3
pyparsing>=3.2.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2025.1
pyxnat>=1.6.3
PyYAML>=6.0.2
ragas>=0.2.14
rdflib>=6.3.2
regex>=2024.11.6
requests>=2.32.3
requests-toolbelt>=1.0.0
safetensors>=0.5.3
scipy>=1.15.2
simplejson>=3.20.1
six>=1.17.0
sniffio>=1.3.1
SQLAlchemy>=2.0.38
tenacity<=9.0.0
tiktoken>=0.9.0
tokenizers>=0.12.1
tqdm>=4.67.1
traits>=7.0.2
transformers>=4.49.0
typing-inspect>=0.9.0
typing-extensions>=4.12.2
tzdata>=2025.1
urllib3>=2.3.0
xxhash>=3.5.0
yarl>=1.18.3
zstandard>=0.23.0
